+++ CHANGELOG.md
@@ -7,12 +7,16 @@ All notable changes to this project will be documented in this file.
 ### 🚀 Features
 
 - *(rasterizer)* Add beamterm-rasterizer for glyph rasterization on native targets
-- *(core)* Add feature-gated native dynamic font atlas and rasterizer
+- *(core)* Add feature-gated native dynamic font atlas and rasterizer (#101)
 
 ### 💼 Other
 
 - *(core)* Add a simple terminal-emulator example
 - *(deps)* Bump cosmic-text from 0.16.0 to 0.18.2
+
+### 🚜 Refactor
+
+- *(core)* Eliminate RefCell usage in core
 
 ## [beamterm-v0.17.0] - 2026-03-11